Make course_experience a feature
