+++ lms/static/course_experience
@@ -1,0 +1,1 @@
+../../openedx/features/course_experience/static/course_experience
