+++ openedx/features/README.rst
@@ -1,0 +1,6 @@
+Open EdX Features
+-----------------
+
+This is the root package for Open edX features that extend the edX platform.
+The intention is that these features would ideally live in an external
+repository, but for now they live in edx-platform but are cleanly modularized.
