+++ openedx/features/course_experience/static/course_experience/fixtures/course-outline-fragment.html
