+++ openedx/features/course_experience/templates/course_experience/course-home.html
