+++ openedx/features/course_experience/templates/course_experience/course-outline-fragment.html
